feat: Scotiabank products

+++ lib/models/available_amount.dart
@@ -1,13 +1,13 @@
 class AvailableAmount {
   final String currency;
-  final double amount;
+  final int amount;
 
   AvailableAmount({required this.currency, required this.amount});
 
   factory AvailableAmount.fromJson(Map<String, dynamic> json) {
     return AvailableAmount(
       currency: json['currency'] as String,
-      amount: (json['amount'] as num).toDouble(),
+      amount: (json['amount'] as num).toInt(),
     );
   }
 
@@ -15,6 +15,3 @@ class AvailableAmount {
     return {'currency': currency, 'amount': amount};
   }
 }
-
-
-
